Improve TextPartView performance.
Most importantly: never access db on every message update, everything has to be
prefetched.
Second: it's better to create separate signals for each property than to create
one big signal and later pattern match on it's result. Smaller signals are
faster, will often not require thread switch, and will not block each other.
It's also important to cut signal processing as early as possible, for example
there is no need to fetch normalized text message content if we are not in
search result mode (content search query is empty).

Added signal in CollectionController containing ids of messages matching current
text search (present only when query is not empty). This set is prefetched only
when needed and TextPartView can easily check if needs highlighting.
In most cases this won't be needed so we can skip many operations.

Removed calls to fetch normalized content from db, this is slow, as index
storage is not cached, also index storage contains content for whole message,
not for specific parts. We can easily compute normalized content when needed,
it should be fast enough.